Move phone onboarding steps into separate files
