+++ frontend/src/components/phones/onboarding/RealPhoneSetup.tsx
@@ -1,0 +1,318 @@
+import styles from "./PhoneOnboarding.module.scss";
+import { Localized, useLocalization } from "@fluent/react";
+import PhoneVerify from "./images/phone-verify.svg";
+import EnteryVerifyCode from "./images/enter-verify-code.svg";
+import EnteryVerifyCodeError from "./images/verify-code-error.svg";
+import { Button } from "../../Button";
+import {
+  hasPendingVerification,
+  UnverifiedPhone,
+  VerificationPendingPhone,
+  PhoneNumberSubmitVerificationFn,
+} from "../../../hooks/api/phone";
+import {
+  ChangeEventHandler,
+  FormEventHandler,
+  useEffect,
+  useRef,
+  useState,
+} from "react";
+import { parseDate } from "../../../functions/parseDate";
+
+type RealPhoneSetupProps = {
+  unverifiedRealPhones: Array<UnverifiedPhone | VerificationPendingPhone>;
+  onRequestVerification: (numberToVerify: string) => void;
+  onSubmitVerification: PhoneNumberSubmitVerificationFn;
+};
+export const RealPhoneSetup = (props: RealPhoneSetupProps) => {
+  const phonesPendingVerification = props.unverifiedRealPhones.filter(
+    hasPendingVerification
+  );
+  const [isEnteringNumber, setIsEnteringNumber] = useState(
+    phonesPendingVerification.length === 0
+  );
+
+  if (isEnteringNumber || phonesPendingVerification.length === 0) {
+    return (
+      <RealPhoneForm
+        requestPhoneVerification={(number) => {
+          props.onRequestVerification(number);
+          setIsEnteringNumber(false);
+        }}
+      />
+    );
+  }
+
+  return (
+    <RealPhoneVerification
+      phonesPendingVerification={phonesPendingVerification}
+      submitPhoneVerification={props.onSubmitVerification}
+      onGoBack={() => setIsEnteringNumber(true)}
+    />
+  );
+};
+type RealPhoneFormProps = {
+  requestPhoneVerification: (phoneNumber: string) => void;
+};
+const RealPhoneForm = (props: RealPhoneFormProps) => {
+  const { l10n } = useLocalization();
+
+  const [phoneNumber, setPhoneNumber] = useState("");
+
+  const errorMessage = (
+    <div className={`${styles["error"]}`}>
+      {l10n.getString("phone-onboarding-step2-invalid-number", {
+        phone_number: "+1 (000) 000-0000",
+      })}
+    </div>
+  );
+
+  const onChange: ChangeEventHandler<HTMLInputElement> = (event) => {
+    setPhoneNumber(event.target.value);
+  };
+
+  const onSubmit: FormEventHandler = (event) => {
+    event.preventDefault();
+    props.requestPhoneVerification(phoneNumber);
+  };
+
+  return (
+    <div className={`${styles.step}  ${styles["step-verify-input"]} `}>
+      <div className={styles.lead}>
+        <img src={PhoneVerify.src} alt="" width={200} />
+        <h2>{l10n.getString("phone-onboarding-step2-headline")}</h2>
+        <p>{l10n.getString("phone-onboarding-step2-body")}</p>
+      </div>
+      {/* Add error class of `mzp-is-error` */}
+      <form onSubmit={onSubmit} className={styles.form}>
+        {/* TODO: Wrap error message in logic */}
+        {errorMessage}
+
+        <input
+          className={`${styles["c-verify-phone-input"]}`}
+          placeholder={l10n.getString(
+            "phone-onboarding-step2-input-placeholder"
+          )}
+          type="tel"
+          required={true}
+          onChange={onChange}
+          autoFocus={true}
+        />
+        {/* TODO: Add error class to input field */}
+        <Button className={styles.button} type="submit">
+          {l10n.getString("phone-onboarding-step2-button-cta")}
+        </Button>
+      </form>
+    </div>
+  );
+};
+// TODO: Add logic to display the correct information between: default/error/success for the following Step Three
+// Init state: Enter verification
+// If 5 mins expire, show errorTimeExpired message
+// If code is wrong, add is-error classes and update image/title
+// If code is correct, show successWhatsNext and  update image/title
+type IntervalFunction = () => unknown | void;
+function useInterval(callback: IntervalFunction, delay: number) {
+  const savedCallback = useRef<IntervalFunction | null>(null);
+
+  // Remember the latest callback.
+  useEffect(() => {
+    savedCallback.current = callback;
+  }, [callback]);
+
+  // Set up the interval.
+  useEffect(() => {
+    function tick() {
+      if (savedCallback.current !== null) {
+        savedCallback.current();
+      }
+    }
+    if (delay !== null) {
+      const id = setInterval(tick, delay);
+      return () => clearInterval(id);
+    }
+  }, [delay]);
+}
+function getRemainingTime(verificationSentTime: Date): number {
+  const currentTime = new Date().getTime();
+  const elapsedTime = currentTime - verificationSentTime.getTime();
+  const remainingTime = 5 * 60 * 1000 - elapsedTime;
+  return remainingTime;
+}
+function formatTime(remainingMilliseconds: number) {
+  const remainingSeconds = Math.floor(remainingMilliseconds / 1000);
+  const minutes = Math.floor(remainingSeconds / 60);
+  const seconds = remainingSeconds - minutes * 60;
+
+  // Add a leading 0 if it's under 10 seconds.
+  const formmatedSeconds = seconds < 10 ? `0${seconds}` : seconds;
+
+  return `${minutes}:${formmatedSeconds}`;
+}
+function formatPhone(phoneNumber: string) {
+  // Bug: Any country code with two characters will break
+  // Return in +1 (111) 111-1111 format
+  return `${phoneNumber.substring(0, 2)} (${phoneNumber.substring(
+    2,
+    5
+  )}) ${phoneNumber.substring(5, 8)}-${phoneNumber.substring(8, 12)}`;
+}
+function getPhoneWithMostRecentlySentVerificationCode(
+  phones: VerificationPendingPhone[]
+) {
+  const mostRecentVerifiedPhone = [...phones].sort((a, b) => {
+    const sendDateA = parseDate(a.verification_sent_date).getTime();
+    const sendDateB = parseDate(b.verification_sent_date).getTime();
+    return sendDateA - sendDateB;
+  });
+
+  return mostRecentVerifiedPhone[0];
+}
+type RealPhoneVerificationProps = {
+  phonesPendingVerification: VerificationPendingPhone[];
+  submitPhoneVerification: PhoneNumberSubmitVerificationFn;
+  onGoBack: () => void;
+};
+const RealPhoneVerification = (props: RealPhoneVerificationProps) => {
+  const { l10n } = useLocalization();
+
+  const inputRef = useRef<HTMLInputElement>(null);
+  const leadRef = useRef<HTMLDivElement>(null);
+  const expiredErrorMsg = useRef<HTMLDivElement>(null);
+  const formRef = useRef<HTMLFormElement>(null);
+  const phoneWithMostRecentlySentVerificationCode =
+    getPhoneWithMostRecentlySentVerificationCode(
+      props.phonesPendingVerification
+    );
+  const [verificationCode, setVerificationCode] = useState("");
+  const verificationSentDate = parseDate(
+    phoneWithMostRecentlySentVerificationCode.verification_sent_date
+  );
+  const [remainingTime, setRemainingTime] = useState(
+    getRemainingTime(verificationSentDate)
+  );
+
+  const onChange: ChangeEventHandler<HTMLInputElement> = (event) => {
+    setVerificationCode(event.target.value);
+  };
+
+  const onSubmit: FormEventHandler = async (event) => {
+    event.preventDefault();
+
+    const response = await props.submitPhoneVerification(
+      phoneWithMostRecentlySentVerificationCode.id,
+      {
+        number: phoneWithMostRecentlySentVerificationCode.number,
+        verification_code: verificationCode,
+      }
+    );
+
+    if (response.ok) {
+      leadRef.current?.classList.add(`${styles["is-success"]}`);
+    } else {
+      leadRef.current?.classList.add(`${styles["is-error"]}`);
+      inputRef.current?.classList.add(`${styles["is-error"]}`);
+    }
+  };
+
+  useInterval(() => {
+    if (remainingTime < 0) {
+      expiredErrorMsg.current?.classList.remove(`${styles["is-hidden"]}`);
+      formRef.current?.classList.add(`${styles["is-hidden"]}`);
+    }
+
+    setRemainingTime(getRemainingTime(verificationSentDate));
+  }, 1000);
+
+  const errorTimeExpired = (
+    <div
+      ref={expiredErrorMsg}
+      className={`${styles["step-input-verificiation-code-timeout"]} ${styles["is-hidden"]} `}
+    >
+      <p>{l10n.getString("phone-onboarding-step3-error-exipred")}</p>
+      <Button className={styles.button} type="submit">
+        {l10n.getString("phone-onboarding-step3-error-cta")}
+      </Button>
+    </div>
+  );
+
+  const codeEntryPlaceholder = "000000";
+
+  return (
+    <div
+      className={`${styles.step}  ${styles["step-input-verificiation-code"]} `}
+    >
+      <div ref={leadRef} className={styles.lead}>
+        <div
+          className={`${styles["step-input-verificiation-code-lead-default"]}`}
+        >
+          {/* Default state */}
+          <img src={EnteryVerifyCode.src} alt="" width={300} />
+          <h2>{l10n.getString("phone-onboarding-step2-headline")}</h2>
+        </div>
+        <div
+          className={`${styles["step-input-verificiation-code-lead-error"]} `}
+        >
+          {/* Timeout error state */}
+          <img src={EnteryVerifyCodeError.src} alt="" width={170} />
+          <h2 className={`${styles["is-error"]} `}>
+            {l10n.getString("phone-onboarding-step3-code-fail-title")}
+          </h2>
+          <p>{l10n.getString("phone-onboarding-step3-code-fail-body")}</p>
+        </div>
+      </div>
+      {/* Add error class of `mzp-is-error` */}
+
+      {/* TODO: Add logic to display timeout error */}
+      {errorTimeExpired}
+
+      {/* <form onSubmit={onSubmit} className={`${styles.form}`}> */}
+      <form ref={formRef} onSubmit={onSubmit} className={styles.form}>
+        {/* TODO: Make remaining_time count backwards with "X minutes, XX seconds" */}
+        <Localized
+          id="phone-onboarding-step3-body"
+          vars={{
+            phone_number: formatPhone(
+              phoneWithMostRecentlySentVerificationCode.number
+            ),
+            remaining_time: formatTime(remainingTime),
+          }}
+          elems={{
+            span: <span className={`${styles["phone-number"]}`} />,
+            strong: <strong />,
+          }}
+        >
+          <p />
+        </Localized>
+
+        <input
+          placeholder={codeEntryPlaceholder}
+          required={true}
+          onChange={onChange}
+          ref={inputRef}
+          autoFocus={true}
+        />
+        {/* TODO: Add logic to show success/fail on submit */}
+        <Button className={styles.button} type="submit">
+          {l10n.getString("phone-onboarding-step3-button-cta")}
+        </Button>
+
+        {/* TODO: Enable logic to "go back" to previous step */}
+        <Button
+          onClick={() => props.onGoBack()}
+          className={styles.button}
+          type="button"
+          variant="secondary"
+        >
+          {l10n.getString("phone-onboarding-step3-button-edit")}
+        </Button>
+      </form>
+
+      {/* TODO: Resubmit phone number for verification and reset count down */}
+      {/* onClick={() => props.requestPhoneVerification(phoneNumber)} */}
+      <Button className={styles.button} type="button" variant="secondary">
+        {l10n.getString("phone-onboarding-step3-button-resend")}
+      </Button>
+    </div>
+  );
+};
